fix(planner): an integer CAST keeps its bigint declaration, measured
#1070 narrowed four sites to int4. Three of them — an integer LITERAL,
int4 arithmetic and unary minus — declare the same thing on the single
path and on all three stage arms. The CAST does not: the DAG declares a
cast OVER A WINDOW from the window's own output rather than from the
cast, so `CAST(SUM(a) OVER () AS INTEGER)` came back int4 on the single
path and int8 on both DAG arms. One expression, two declarations, which
is exactly what #813 item 1 was and what this arc exists to prevent.

Measured, not guessed: `CAST(id AS INTEGER)` and a cast over a GROUPED
aggregate agree on both paths, and so do `MIN(id) OVER () + 1`,
`-MIN(id) OVER ()`, `BITWISE_AND(MIN(id) OVER (), 3)` and a bare literal.
The window is the one producer whose absorbed projection is typed
somewhere this walk does not reach, and finding that site is its own
change.

So inferCastType keeps INT64 for every integer spelling and the
divergence stays in ADR-0012's list, where it already was. The six
`int4/*` pins in pgwire.TestScalarSubqueryAggregateMatrix keep their
pins for the same reason; the six `max/*` pins beside them are still
gone, because their int4 comes from an aggregate over an int4 COLUMN.

+++ wadjet/unknown_cast_type_test.go
@@ -86,10 +86,10 @@ func TestUnknownCastDestinationIsUndefinedObject(t *testing.T) {
 		want      any
 	}{
 		{"bigint", `SELECT CAST(c_i64 AS BIGINT) AS v FROM ` + tbl + ` WHERE id = 1`, int64(1000003)},
-		{"integer", `SELECT CAST(c_i32 AS INTEGER) AS v FROM ` + tbl + ` WHERE id = 1`, int32(3)},
-		{"smallint", `SELECT CAST(c_i32 AS SMALLINT) AS v FROM ` + tbl + ` WHERE id = 1`, int32(3)},
-		{"int2", `SELECT CAST(c_i32 AS INT2) AS v FROM ` + tbl + ` WHERE id = 1`, int32(3)},
-		{"int4", `SELECT CAST(c_i32 AS INT4) AS v FROM ` + tbl + ` WHERE id = 1`, int32(3)},
+		{"integer", `SELECT CAST(c_i32 AS INTEGER) AS v FROM ` + tbl + ` WHERE id = 1`, int64(3)},
+		{"smallint", `SELECT CAST(c_i32 AS SMALLINT) AS v FROM ` + tbl + ` WHERE id = 1`, int64(3)},
+		{"int2", `SELECT CAST(c_i32 AS INT2) AS v FROM ` + tbl + ` WHERE id = 1`, int64(3)},
+		{"int4", `SELECT CAST(c_i32 AS INT4) AS v FROM ` + tbl + ` WHERE id = 1`, int64(3)},
 		{"signed", `SELECT CAST(c_i32 AS SIGNED) AS v FROM ` + tbl + ` WHERE id = 1`, int64(3)},
 		{"real", `SELECT CAST(1.0/3 AS REAL) AS v FROM ` + tbl + ` WHERE id = 1`, float32(0.33333334)},
 		{"float4", `SELECT CAST(1.0/3 AS FLOAT4) AS v FROM ` + tbl + ` WHERE id = 1`, float32(0.33333334)},
